% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairWQS_functions.R
\name{pairwqs}
\alias{pairwqs}
\title{Weighted Quantile Sum Regression for design with paired data structure}
\usage{
pairwqs(
  train_data,
  valid_data = NULL,
  col_vars,
  col_covars,
  id = "studyid",
  event = "event",
  q = 4,
  boot = FALSE,
  B = 10
)
}
\arguments{
\item{train_data}{The training data}

\item{valid_data}{The validation data}

\item{col_vars}{Exposure variable names}

\item{col_covars}{Covariates}

\item{id}{Matching ID variable}

\item{event}{Event indicator}

\item{q}{Number of quantiles}

\item{boot}{Whether to perform bootstrap}

\item{B}{Number of bootstrap}
}
\value{
A list containing WQS weights and the regression coefficient of WQS index
}
\description{
Weighted Quantile Sum Regression for design with paired data structure
}
