% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairWQS_functions.R
\name{pairwqs_boot}
\alias{pairwqs_boot}
\title{Weighted Quantile Sum Regression with Bootstrapping}
\usage{
pairwqs_boot(
  wqsdata,
  col_vars,
  col_covars,
  id = "studyid",
  event = "event",
  q = 4,
  B = 10
)
}
\arguments{
\item{wqsdata}{The input data}

\item{col_vars}{Exposure variable names}

\item{col_covars}{Covariates}

\item{id}{Matching ID variable}

\item{event}{Event indicator}

\item{q}{Number of quantiles}

\item{B}{Number of bootstrap}
}
\value{
A list containing WQS weights and the regression coefficient of WQS index
}
\description{
Weighted Quantile Sum Regression with Bootstrapping
}
