% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairWQS_functions.R
\name{pairwqs_noboot}
\alias{pairwqs_noboot}
\title{Weighted Quantile Sum Regression without Bootstrapping}
\usage{
pairwqs_noboot(
  wqsdata,
  col_vars,
  col_covars,
  id = "studyid",
  event = "event",
  q = 4
)
}
\arguments{
\item{wqsdata}{The input data}

\item{col_vars}{Exposure variable names}

\item{col_covars}{Covariates}

\item{id}{Matching ID variable}

\item{event}{Event indicator}

\item{q}{Number of quantiles}
}
\value{
A list containing WQS weights and the regression coefficient of WQS index
}
\description{
Weighted Quantile Sum Regression without Bootstrapping
}
